requirements.txt
psycopg2-binary>=2.6.1
pyyaml>=3.11
tqdm>=1.0
stop-words>=2015.2.23.1
docopt>=0.6.1
numpy>=1.9.2
pathlib
